requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
stripe==7.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9